Run EM to estimate parameters. Iteratively run susie and estimate parameters - RSS version

Run EM to estimate parameters. Iteratively run susie and estimate parameters - RSS version

EM(
  region_data,
  niter = 20,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_all", "shared_type"),
  use_null_weight = TRUE,
  max_iter = 1,
  ncore = 1,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object with the susie input data for each region

niter

the number of iterations of the E-M algorithm to perform

init_group_prior

a vector of initial prior inclusion probabilities for SNPs and genes.

init_group_prior_var

a vector of initial prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "independent" is the default and allows all groups to have their own separate variance parameters. "shared_all" allows all groups to share the same variance parameter. "shared_type" allows all groups in one molecular QTL type to share the same variance parameter.

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

max_iter

Maximum number of IBSS iterations to perform.

ncore

The number of cores used to parallelize susie over regions

verbose

TRUE/FALSE. If TRUE, print detail messages

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies.

Value

a list of parameters