Summarize and plot cTWAS parameter estimates

Summarize and plot cTWAS parameter estimates

ctwas_summarize_parameters(
  outputdir,
  outname,
  gwas_n = NA,
  thin = 1,
  plot_estimates = T
)

Arguments

outputdir

a string, the directory where ctwas output is stored

outname

a string, the output name

gwas_n

the sample size of the GWAS summary statistics

thin

the proportion of SNPs used for parameter estimation, as supplied to ctwas_rss

plot_estimates

if TRUE, return a plot of the estimated parameters