Causal inference for TWAS using summary statistics

Causal inference for TWAS using summary statistics

ctwas_sumstats(
  z_snp,
  weights,
  region_info,
  z_gene = NULL,
  thin = 0.1,
  niter_prefit = 3,
  niter = 30,
  L = 5,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_type", "shared_context", "shared_nonSNP",
    "shared_all", "independent"),
  maxSNP = 20000,
  min_nonSNP_PIP = 0.5,
  p_single_effect = 0.8,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  save_cor = FALSE,
  cor_dir = NULL,
  outputdir = NULL,
  outname = "ctwas",
  logfile = NULL,
  verbose = FALSE,
  ...
)

Arguments

z_snp

A data frame with four columns: "id", "A1", "A2", "z". giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.

weights

a list of weights

region_info

a data frame of region definition and associated LD file names

z_gene

A data frame with columns: "id", "z", giving the z-scores for genes.

thin

The proportion of SNPs to be used for estimating parameters and screening regions.

niter_prefit

the number of iterations of the E-M algorithm to perform during the initial parameter estimation step

niter

the number of iterations of the E-M algorithm to perform during the complete parameter estimation step

L

the number of effects for susie during the fine mapping steps

init_group_prior

a vector of initial values of prior inclusion probabilities for SNPs and genes.

init_group_prior_var

a vector of initial values of prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "shared_type" allows all groups in one molecular QTL type to share the same variance parameter. "shared_context" allows all groups in one molecular QTL type to share the same variance parameter. "shared_nonSNP" allows all non-SNP groups to share the same variance parameter. "shared_all" allows all groups to share the same variance parameter. "independent" allows all groups to have their own separate variance parameters.

maxSNP

Inf or integer. Maximum number of SNPs in a region. Default is Inf, no limit. This can be useful if there are many SNPs in a region and you don't have enough memory to run the program. This applies to the finemapping step only.

min_nonSNP_PIP

Regions with non-SNP PIP >= min_nonSNP_PIP will be selected to run finemapping using full SNPs.

p_single_effect

Regions with probability greater than p_single_effect of having 1 or fewer effects will be used for parameter estimation

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a . credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies.

ncore

The number of cores used to parallelize susie over regions

save_cor

TRUE/FALSE. If TRUE, save correlation (R) matrices to cor_dir

cor_dir

The directory to store correlation (R) matrices

outputdir

The directory to store output. If specified, save outputs to the directory.

outname

The output name.

logfile

path to the log file, if NULL will print log info on screen.

verbose

TRUE/FALSE. If TRUE, print detailed messages

...

Additional arguments of susie_rss.

Value

a list of estimated parameters, fine-mapping results, boundary genes, z_gene, and region_data