Estimate cTWAS parameters

Estimate cTWAS parameters

est_param(
  z_snp,
  region_info,
  regionlist = NULL,
  z_gene = NULL,
  gene_info = NULL,
  weights = NULL,
  thin = 1,
  prob_single = 0.8,
  niter1 = 3,
  niter2 = 30,
  group_prior = NULL,
  group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_all", "shared+snps",
    "shared_QTLtype"),
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  logfile = NULL
)

Arguments

z_snp

a data frame with four columns: "id", "A1", "A2", "z". giving the z scores for SNPs. "A1" is effect allele. "A2" is the other allele.

region_info

a data frame of region definition and associated file names

regionlist

a list object indexing regions, variants and genes.

z_gene

a data frame with two columns: "id", "z". giving the z scores for genes. Optionally, a "type" column can also be supplied; this is for using multiple sets of weights

gene_info

a data frame of gene information obtained from compute_gene_z

thin

The proportion of SNPs to be used for the parameter estimation and initial screening region steps. Smaller thin parameters reduce runtime at the expense of accuracy. The fine mapping step is rerun using full SNPs for regions with strong gene signals.

prob_single

Blocks with probability greater than prob_single of having 1 or fewer effects will be used for parameter estimation

niter1

the number of iterations of the E-M algorithm to perform during the initial parameter estimation step

niter2

the number of iterations of the E-M algorithm to perform during the complete parameter estimation step

group_prior

a vector of two prior inclusion probabilities for SNPs and genes.

group_prior_var

a vector of two prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "independent" is the default and allows all groups to have their own separate variance parameters. "shared" allows all groups to share the same variance parameter. "shared+snps" allows all groups to share the same variance parameter, and this variance parameter is also shared with SNPs. "shared_type" allows all groups in one molecular QTL type to share the same variance parameter.

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies.

ncore

The number of cores used to parallelize computation over regions

logfile

the log file, if NULL will print log info on screen

weight_list

a list of weights by chromosome

Value

a list with estimated parameters, and updated region_info with correlation file names for each region.