Estimate L for all regions by running finemapping with uniform prior

Estimate L for all regions by running finemapping with uniform prior

estimate_region_L(
  region_data,
  LD_map,
  weights,
  init_L = 5,
  min_abs_corr = 0.1,
  snps_only = FALSE,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  ncore = 1,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object indexing regions, variants and genes.

LD_map

a data frame with filenames of LD matrices for each of the regions.

weights

a list of preprocessed weights.

init_L

upper bound of the number of causal signals

min_abs_corr

Minimum absolute correlation allowed in a credible set.

snps_only

If TRUE, use only SNPs in the region data.

LD_format

file format for LD matrix. If "custom", use a user defined LD_loader_fun() function to load LD matrix.

LD_loader_fun

a user defined function to load LD matrix when LD_format = "custom".

snpinfo_loader_fun

a user defined function to load SNP information file, if SNP information files are not in standard cTWAS reference format.

ncore

The number of cores used to parallelize susie over regions

verbose

If TRUE, print detail messages

...

Additional arguments of susie_rss.

Value

a vector of estimated L for all regions