Runs cTWAS finemapping for a single region

Runs cTWAS finemapping for a single region

finemap_region(
  region_data,
  region_id,
  use_LD = TRUE,
  LD_info = NULL,
  snp_info = NULL,
  weights = NULL,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  LD_format = c("rds", "rdata", "csv", "txt", "custom"),
  LD_loader = NULL,
  include_cs_index = TRUE,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object with data for the regions

region_id

a character string of region id to be finemapped

use_LD

TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.

LD_info

a list of paths to LD matrices for each of the regions. Required when use_LD = TRUE.

snp_info

a list of SNP info data frames for LD reference. Required when use_LD = TRUE.

weights

a list of preprocessed weights

L

the number of effects for susie during the fine mapping

group_prior

a vector of two prior inclusion probabilities for SNPs and genes.

group_prior_var

a vector of two prior variances for SNPs and gene effects.

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies.

force_compute_cor

TRUE/FALSE. If TRUE, force computing correlation (R) matrices

save_cor

TRUE/FALSE. If TRUE, save correlation (R) matrices to cor_dir

cor_dir

a string, the directory to store correlation (R) matrices

LD_format

file format for LD matrix. If "custom", use a user defined LD_loader() function to load LD matrix.

LD_loader

a user defined function to load LD matrix when LD_format = "custom".

include_cs_index

TRUE/FALSE. If TRUE, add cs_index to finemapping results.

verbose

TRUE/FALSE. If TRUE, print detail messages

...

Additional arguments of susie_rss.

Value

a data frame of finemapping results.