Runs cTWAS fine-mapping for regions

Runs cTWAS fine-mapping for regions

finemap_regions(
  region_data,
  LD_map,
  weights,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.1,
  include_cs = TRUE,
  get_susie_alpha = TRUE,
  snps_only = FALSE,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)

Arguments

region_data

region_data to be finemapped

LD_map

a data frame with filenames of LD matrices for the regions.

weights

a list of preprocessed weights.

L

the number of effects or a vector of number of effects for each region.

group_prior

a vector of two prior inclusion probabilities for SNPs and genes. If NULL, it will use uniform prior inclusion probabilities.

group_prior_var

a vector of two prior variances for SNPs and gene effects. If NULL, it will set prior variance = 50 as the default in susie_rss.

use_null_weight

If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a credible set.

include_cs

If TRUE, add credible sets (CS) to finemapping results.

get_susie_alpha

If TRUE, get susie alpha matrix from finemapping results.

snps_only

If TRUE, use only SNPs in the region data.

force_compute_cor

If TRUE, force computing correlation (R) matrices

save_cor

If TRUE, save correlation (R) matrices to cor_dir

cor_dir

a string, the directory to store correlation (R) matrices

LD_format

file format for LD matrix. If "custom", use a user defined LD_loader_fun() function to load LD matrix.

LD_loader_fun

a user defined function to load LD matrix when LD_format = "custom".

snpinfo_loader_fun

a user defined function to load SNP information file, if SNP information files are not in standard cTWAS reference format.

ncore

The number of cores used to parallelize computation over regions

verbose

If TRUE, print detail messages

logfile

the log file, if NULL will print log info on screen

...

Additional arguments of susie_rss.

Value

a list with cTWAS finemapping results.