Runs cTWAS finemapping for a single region without LD

Runs cTWAS finemapping for a single region without LD

finemap_single_region_noLD(
  region_data,
  region_id,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  include_cs = TRUE,
  get_susie_alpha = TRUE,
  snps_only = FALSE,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object with data for the regions

region_id

a character string of region id to be finemapped

group_prior

a vector of two prior inclusion probabilities for SNPs and genes. If NULL, it will use uniform prior inclusion probabilities.

group_prior_var

a vector of two prior variances for SNPs and gene effects. If NULL, it will set prior variance = 50 as the default in susie_rss.

use_null_weight

If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

include_cs

If TRUE, add credible sets (CS) to finemapping results.

get_susie_alpha

If TRUE, get susie alpha matrix from finemapping results.

snps_only

If TRUE, use only SNPs in the region data.

verbose

If TRUE, print detail messages

...

Additional arguments of susie_rss.

Value

a data frame of finemapping results.