Impute expression

Impute expression

impute_expr(
  pgenfs,
  weight,
  method = "lasso",
  outputdir = getwd(),
  outname = NULL,
  logfile = NULL,
  compress = T,
  harmonize = T
)

Arguments

pgenfs

A character vector of .pgen or .bed files. One file for one chromosome, in the order of 1 to 22. Therefore, the length of this vector needs to be 22. If .pgen files are given, then .pvar and .psam are assumed to present in the same directory. If .bed files are given, then .bim and .fam files are assumed to present in the same directory.

weight

a string, pointing to a directory with the fusion/twas format of weights, or a vector of one or more .db files in PredictDB format.

method

a string, blup/bslmm/lasso/top1/enet/best. This option is only used for fusion weights. "best" means the method giving the best cross #' validation R^2. Note that top1 uses only the weight with largest effect.

outputdir

a string, the directory to store output

outname

a string, the output name

logfile

the log file, if NULL will print log info on screen

compress

TRUE/FALSE. If TRUE, the imputed expression files are compressed

harmonize

T/F, if need to harmonize SNP data, if T, will harmonize GWAS and eQTL genotype alleles.