Loads precomputed correlation matrices for a single region.

It loads precomputed correlation matrices for a single region. It could load correlation matrices by region_id and directory of correlation matrices cor_dir. Otherwise, it loads correlation matrices by the filenames (R_sg_file, R_sg_file, R_s_file) if they are provided.

Usage

load_region_cor(region_id, cor_dir, R_sg_file, R_g_file, R_s_file)

Arguments

region_id

a character string of region id.

cor_dir

a string, the directory to store correlation matrices.

R_sg_file

filename of SNP-gene correlations.

R_g_file

filename of gene-gene correlations.

R_s_file

filename of SNP-SNP correlations.

Value

correlation matrices (R_snp, R_snp_gene and R_gene)