Rerun cTWAS finemapping using summary statistics with L = 1 for problematic regions

Rerun cTWAS finemapping using summary statistics with L = 1 for problematic regions

rerun_finemap_regions_L1(
  z_snp,
  z_gene,
  weight,
  problematic_snps = NULL,
  rerun_region_tags = NULL,
  pip_thresh = 0.5,
  ctwas_outputdir = NULL,
  ctwas_outname = NULL,
  ld_R_dir = NULL,
  outputdir = NULL,
  outname = NULL,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = T,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  logfile = NULL
)

Arguments

z_snp

A data frame with four columns: "id", "A1", "A2", "z". giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.

z_gene

A data frame with two columns: "id", "z". giving the z scores for genes. Optionally, a "type" column can also be supplied; this is for using multiple sets of weights

weight

a string, weight filename used in cTWAS

problematic_snps

a character vector of problematic SNP rsIDs

rerun_region_tags

a character vector of region tags to rerun

pip_thresh

Minimum PIP value to select regions

ctwas_outputdir

a string, the directory to cTWAS output

ctwas_outname

a string, the directory to cTWAS output name

outputdir

a string, the directory to store output

outname

a string, the output name

group_prior

a vector of two prior inclusion probabilities for SNPs and genes. This is ignored if estimate_group_prior = T

group_prior_var

a vector of two prior variances for SNPs and gene effects. This is ignored if estimate_group_prior_var = T

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies. #'

ncore

The number of cores used to parallelize over regions

logfile

the log file, if NULL will print log info on screen

LD_R_dir

a string, pointing to a directory containing all LD matrix files and variant information.