Iteratively run susie and estimate parameters

Iteratively run susie and estimate parameters

susieI(
  pgenfs,
  exprfs,
  Y,
  regionlist,
  niter = 20,
  L = 1,
  group_prior = NULL,
  group_prior_var = NULL,
  estimate_group_prior = T,
  estimate_group_prior_var = T,
  use_null_weight = T,
  coverage = 0.95,
  standardize = T,
  ncore = 1,
  outputdir = getwd(),
  outname = NULL,
  report_parameters = T
)

Arguments

pgenfs

A character vector of .pgen or .bed files. One file for one chromosome, in the order of 1 to 22. Therefore, the length of this vector needs to be 22. If .pgen files are given, then .pvar and .psam are assumed to present in the same directory. If .bed files are given, then .bim and .fam files are assumed to present in the same directory.

Y

a vector of length n, phenotype, the same order as provided by `.pgenfs` (defined in .psam or .fam files).

regionlist

a list object indexing regions, variants and genes. The output of index_regions

niter

the number of iterations of the E-M algorithm to perform

L

the number of effects for susie

group_prior

a vector of two prior inclusion probabilities for SNPs and genes. This is ignored if estimate_group_prior = T

group_prior_var

a vector of two prior variances for SNPs and gene effects. This is ignored if estimate_group_prior_var = T

estimate_group_prior

TRUE/FALSE. If TRUE, the prior inclusion probabilities for SNPs and genes are estimated using the data. If FALSE, group_prior must be specified

estimate_group_prior_var

TRUE/FALSE. If TRUE, the prior variances for SNPs and genes are estimated using the data. If FALSE, group_prior_var must be specified

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

standardize

TRUE/FALSE. If TRUE, all variables are standardized to unit variance

ncore

The number of cores used to parallelize susie over regions

outputdir

a string, the directory to store output

outname

a string, the output name

report_parameters

TRUE/FALSE. If TRUE, estimated parameters are reported at the end of iteration

A

character vector of .`expr` or `.expr.gz` files. One file for one chromosome, in the order of 1 to 22. Therefore, the length of this vector needs to be 22. `.expr.gz` file is gzip compressed `.expr` files. `.expr` is a matrix of imputed expression values, row is for each sample, column is for each gene. Its sample order is same as in files provided by `.pgenfs`. We also assume corresponding `.exprvar` files are present in the same directory. `.exprvar` files are just tab delimited text files, with columns:

chrom

chromosome number, numeric

p0

gene boundary position, the smaller value

p1

gene boundary position, the larger value

id

gene id

Its rows should be in the same order as the columns for corresponding `.expr` files.