Causal inference for TWAS

Causal inference for TWAS

ctwas(pgenfs, exprfs, Y, ld_regions = c("EUR", "ASN", "AFR"),
  ld_regions_custom = NULL, thin = 1, prob_single = 0.8,
  niter1 = 3, niter2 = 30, L = 5, group_prior = NULL,
  group_prior_var = NULL, estimate_group_prior = T,
  estimate_group_prior_var = T, use_null_weight = T, coverage = 0.95,
  stardardize = T, ncore = 1, outputdir = getwd(), outname = NULL,
  logfile = NULL)

Arguments

pgenfs
A character vector of .pgen or .bed files. One file for one chromosome, in the order of 1 to 22. Therefore, the length of this vector needs to be 22. If .pgen files are given, then .pvar and .psam are assumed to present in the same directory. If .bed files are given, then .bim and .fam files are assumed to present in the same directory.

exprfs
A character vector of .`expr` or `.expr.gz` files. One file for one chromosome, in the order of 1 to 22. Therefore, the length of this vector needs to be 22. `.expr.gz` file is gzip compressed `.expr` files. `.expr` is a matrix of imputed expression values, row is for each sample, column is for each gene. Its sample order is same as in files provided by `.pgenfs`. We also assume corresponding `.exprvar` files are present in the same directory. `.exprvar` files are just tab delimited text files, with columns:

chrom

chromosome number, numeric

p0

gene boundary position, the smaller value

p1

gene boundary position, the larger value

id

gene id

Its rows should be in the same order as the columns for corresponding `.expr` files.

Y
a vector of length n, phenotype, the same order as provided by `.pgenfs` (defined in .psam or .fam files).

ld_regions
A string representing the population to use for defining LD regions. These LD regions are defined by ldetect. The user can also provide custom LD regions matching genotype data, see ld_regions_custom.

ld_regions_custom
A bed format file defining LD regions. The default is NULL; when specified, ld_regions will be ignored.

prob_single
blocks with probility more than this value will be used for parameter estimation

outname
a string, the output name

logfile
the log file, if NULL will print log info on screen