Causal inference for TWAS using summary statistics

Causal inference for TWAS using summary statistics

ctwas_sumstats(
  z_snp,
  weights,
  region_info,
  thin = 1,
  niter1 = 3,
  niter2 = 30,
  L = 5,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_all", "shared_QTLtype"),
  max_snp_region = Inf,
  min_nonSNP_PIP = 0.5,
  p_single_effect = 0.8,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  save_cor = FALSE,
  cor_dir = getwd(),
  logfile = NULL,
  verbose = FALSE,
  ...
)

Arguments

z_snp

A data frame with four columns: "id", "A1", "A2", "z". giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.

weights

a list of weights

region_info

a data frame of region definition and associated LD file names

thin

The proportion of SNPs to be used for parameter estimation and initial screening regions. Smaller thin parameters reduce runtime at the expense of accuracy. The fine mapping step is rerun using full SNPs for regions with strong gene signals; see min_nonSNP_PIP.

niter1

the number of iterations of the E-M algorithm to perform during the initial parameter estimation step

niter2

the number of iterations of the E-M algorithm to perform during the complete parameter estimation step

L

the number of effects for susie during the fine mapping steps

init_group_prior

a vector of initial values of prior inclusion probabilities for SNPs and genes.

init_group_prior_var

a vector of initial values of prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "independent" is the default and allows all groups to have their own separate variance parameters. "shared_all" allows all groups to share the same variance parameter. "shared_QTLtype" allows all groups in one molecular QTL type to share the same variance parameter.

max_snp_region

Inf or integer. Maximum number of SNPs in a region. Default is Inf, no limit. This can be useful if there are many SNPs in a region and you don't have enough memory to run the program. This applies to the last rerun step (using full SNPs and rerun susie for regions with strong gene signals) only.

min_nonSNP_PIP

Regions with non-SNP PIP >= min_nonSNP_PIP will be selected to run finemapping using full SNPs.

p_single_effect

Regions with probability greater than p_single_effect of having 1 or fewer effects will be used for parameter estimation

use_null_weight

TRUE/FALSE. If TRUE, allow for a probability of no effect in susie

coverage

A number between 0 and 1 specifying the “coverage” of the estimated confidence sets

min_abs_corr

Minimum absolute correlation allowed in a . credible set. The default, 0.5, corresponds to a squared correlation of 0.25, which is a commonly used threshold for genotype data in genetic studies.

ncore

The number of cores used to parallelize susie over regions

save_cor

TRUE/FALSE. If TRUE, save correlation (R) matrices to outputdir

cor_dir

a string, the directory to store correlation (R) matrices

logfile

the log file, if NULL will print log info on screen

verbose

TRUE/FALSE. If TRUE, print detail messages

Value

a list of estimated parameters, fine-mapping results, and boundary genes