Estimates cTWAS parameters using EM

Estimates cTWAS parameters using EM

est_param(
  region_data,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_type", "shared_context", "shared_nonSNP",
    "shared_all", "independent"),
  niter_prefit = 3,
  niter = 30,
  min_p_single_effect = 0.8,
  use_null_weight = TRUE,
  min_var = 2,
  min_gene = 1,
  min_group_size = 100,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object indexing regions, variants and genes.

init_group_prior

a vector of initial values of prior inclusion probabilities for SNPs and genes.

init_group_prior_var

a vector of initial values of prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "shared_type" allows all groups in one molecular QTL type to share the same variance parameter. "shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter. "shared_nonSNP" allows all non-SNP groups to share the same variance parameter. "shared_all" allows all groups to share the same variance parameter. "independent" allows all groups to have their own separate variance parameters.

niter_prefit

the number of iterations of the E-M algorithm to perform during the initial parameter estimation step.

niter

the number of iterations of the E-M algorithm to perform during the complete parameter estimation step.

min_p_single_effect

Regions with probability greater than min_p_single_effect of having at most one causal effect will be used selected for the complete parameter estimation step.

use_null_weight

If TRUE, allow for a probability of no effect in susie.

min_var

minimum number of variables (SNPs and genes) in a region.

min_gene

minimum number of genes in a region.

min_group_size

Minimum number of variables in a group.

ncore

The number of cores used to parallelize computation over regions.

logfile

The log filename. If NULL, print log info on screen.

verbose

If TRUE, print detail messages.

...

Additional arguments of susie_rss.

Value

a list with estimated parameters