Expands region_data with all SNPs

Expands region_data with all SNPs

expand_region_data(region_data, snp_map, z_snp, maxSNP = Inf, ncore = 1)

Arguments

region_data

a list of region gene IDs and SNP IDs and associated file names

snp_map

a list of data frames with SNP-to-region map for the reference.

z_snp

A data frame with columns: "id", "z", giving the z-scores for SNPs.

maxSNP

Inf or integer. Maximum number of SNPs in a region. Default is Inf, no limit. This can be useful if there are many SNPs in a region and you don't have enough memory to run the program.

ncore

The number of cores used to parallelize susie over regions

Value

updated region_data with all SNPs