Runs EM to estimate parameters.

Runs EM to estimate parameters.

fit_EM(
  region_data,
  groups,
  types,
  contexts,
  niter = 20,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_type", "shared_context", "shared_nonSNP",
    "shared_all", "independent"),
  use_null_weight = TRUE,
  ncore = 1,
  verbose = FALSE,
  ...
)

Arguments

region_data

a list object with the susie input data for each region

groups

a vector of the groups to estimate parameters

types

a vector of the types to estimate parameters

contexts

a vector of the contexts to estimate parameters

niter

the number of iterations of the E-M algorithm to perform

init_group_prior

a vector of initial prior inclusion probabilities for SNPs and genes.

init_group_prior_var

a vector of initial prior variances for SNPs and gene effects.

group_prior_var_structure

a string indicating the structure to put on the prior variance parameters. "shared_type" allows all groups in one molecular QTL type to share the same variance parameter. "shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter. "shared_nonSNP" allows all non-SNP groups to share the same variance parameter. "shared_all" allows all groups to share the same variance parameter. "independent" allows all groups to have their own separate variance parameters.

use_null_weight

If TRUE, allow for a probability of no effect in susie

ncore

The number of cores used to parallelize susie over regions

verbose

If TRUE, print detail messages

Value

a list of estimated parameters