Bayesian sum of single-effect (SuSiE) linear regression using z scores

Bayesian sum of single-effect (SuSiE) linear regression using z scores

susie_rss(z, R, maf = NULL, maf_thresh = 0, z_ld_weight = 0,
  L = 10, prior_variance = 50, residual_variance = NULL,
  r_tol = 1e-08, prior_weights = NULL, null_weight = NULL,
  estimate_residual_variance = TRUE, estimate_prior_variance = TRUE,
  estimate_prior_method = c("optim", "EM", "simple"),
  check_null_threshold = 0, prior_tol = 1e-09, max_iter = 100,
  s_init = list(), intercept_value = 0, coverage = 0.95,
  min_abs_corr = 0.5, tol = 0.001, verbose = FALSE,
  track_fit = FALSE, check_R = TRUE, check_z = TRUE)

Arguments

z
a p vector of z scores.

R
a p by p symmetric and positive semidefinite correlation matrix. If it is from a reference panel, we recommend a modification on correlation matrix with parameter `z_ld_weight`.

maf
minor allele frequency; to be used along with `maf_thresh` to filter input summary statistics

maf_thresh
variants having MAF smaller than this threshold will be filtered out

z_ld_weight
the weight assigned to the z in the ld matrix, the ld matrix used in model is cov2cor((1-w) R + w zz'). We recomment setting `z_ld_weight` as 1/number of samples in reference panel.

L
maximum number of non-zero effects

prior_variance
the prior variance (vector of length L, or scalar. In latter case gets repeated L times )

residual_variance
the residual variance, a scaler between 0 and 1

r_tol
tolerance level for eigen value check of positive semidefinite matrix of R.

prior_weights
a p vector of prior probability that each element is non-zero

null_weight
probability of no effect, for each single effect model

estimate_residual_variance
indicates whether to estimate residual variance

estimate_prior_variance
indicates whether to estimate prior

estimate_prior_method
The method used for estimating prior variance. "simple" method only compares the loglikelihood between using specified prior variance and using zero, and chose the one that gives larger loglikelihood.

check_null_threshold
when prior variance is estimated, compare the estimate with the null and set prior variance to null (zero) unless the log-likelihood using the estimate is larger than that of null by this threshold. For example, you can set it to 0.1 to nudge the estimate towards zero. Default is 0. Notice that setting it to non-zero may lead to decreasing ELBO in some cases.

prior_tol
when prior variance is estimated, compare the estimated value to this tol at the end of the analysis and exclude a single effect from PIP computation if the estimated prior variance is smaller than it.

max_iter
maximum number of iterations to perform

s_init
a previous susie fit with which to initialize

intercept_value
a value to assign to the intercept (since the intercept cannot be estimated from centered summary data). This value will be used by coef.susie() to assign an intercept value, for consistency with the non-summary-statistic version of this function susie. Set to NULL if you want coef.susie() not to include an intercept term (and so only return a p vector).

coverage
coverage of confident sets. Default to 0.95 for 95% credible interval.

min_abs_corr
minimum of absolute value of correlation allowed in a credible set.

tol
convergence tolerance based on elbo

verbose
if true outputs some progress messages

track_fit
add an attribute trace to output that saves current values of all iterations

check_R
check whether R is positive semidefinite

check_z
check whether z in space spanned by the non-zero eigenvectors of R

Value

a susie fit, which is a list with some or all of the following elements

alpha

an L by p matrix of posterior inclusion probabilites

mu

an L by p matrix of posterior means (conditional on inclusion)

mu2

an L by p matrix of posterior second moments (conditional on inclusion)

Rr

an p vector of t(X) times fitted values, the fitted values equal to X times colSums(alpha*mu))

sigma2

residual variance

V

prior variance

Details

Revised from susieR::susie_rss. Performs sum of single-effect (SuSiE) linear regression with z scores. The summary data required are the p by p correlation matrix R, the p vector z. The summary stats should come from the same individuals. This function fits the regression model z = sum_l Rb_l + e, where e is N(0,residual_variance * R) and the sum_l b_l is a p vector of effects to be estimated. The assumption is that each b_l has exactly one non-zero element, with all elements equally likely to be non-zero. The prior on the non-zero element is N(0,var=prior_variance).

Examples

set.seed(1)
n    <- 1000
p    <- 1000
beta <- rep(0,p)
beta[1:4] <- 1
X        <- matrix(rnorm(n*p),nrow=n,ncol=p)
y        <- c(X %*% beta + rnorm(n))
input_ss <- compute_ss(X,y,standardize = TRUE)
ss <- univariate_regression(X, y)
R <- with(input_ss, cov2cor(XtX))
zhat <- with(ss, betahat/sebetahat)
# res <- susie_rss(zhat, R)